refactor: remove IEverything with a single method
