+++ packages/http/src/index.ts
@@ -1,9 +1,9 @@
 import { factory } from '@stoplight/prism-core';
 import { IHttpOperation } from '@stoplight/types';
 import { defaults } from 'lodash';
-import { mocker } from './mocker';
-import { router } from './router';
-import { validator } from './validator';
+import mock from './mocker';
+import route from './router';
+import { validateInput, validateOutput } from './validator';
 export * from './types';
 import {
   IHttpConfig,
@@ -23,11 +23,7 @@ import {
 const createInstance = (config: IHttpConfig, components?: PickRequired<Partial<PrismHttpComponents>, 'logger'>) => {
   return factory<IHttpOperation, IHttpRequest, IHttpResponse, IHttpConfig>(
     config,
-    defaults(components, {
-      router,
-      validator,
-      mocker,
-    }),
+    defaults(components, { route, validateInput, validateOutput, mock }),
   );
 };
