+++ packages/http/src/validator/__tests__/functional.spec.ts
@@ -1,9 +1,6 @@
 import { DiagnosticSeverity, HttpParamStyles } from '@stoplight/types';
 import { httpInputs, httpOperations, httpOutputs } from '../../__tests__/fixtures';
-import { IHttpConfig } from '../../types';
-import { validator } from '../index';
-
-const defaultConfig: IHttpConfig = { mock: { dynamic: false }, validateRequest: true, validateResponse: true };
+import { validateInput, validateOutput } from '../index';
 
 const BAD_INPUT = Object.assign({}, httpInputs[2], {
   body: { name: 'Shopping', completed: 'yes' },
@@ -23,13 +20,13 @@ const BAD_OUTPUT = Object.assign({}, httpOutputs[1], {
 describe('HttpValidator', () => {
   describe('validateInput()', () => {
     describe('all validations are turned on', () => {
-      it('returns validation errors for whole request structure', async () => {
-        expect(await validator.validateInput({ resource: httpOperations[2], input: BAD_INPUT })).toMatchSnapshot();
+      it('returns validation errors for whole request structure', () => {
+        expect(validateInput({ resource: httpOperations[2], input: BAD_INPUT })).toMatchSnapshot();
       });
 
       describe('when all required params are provided', () => {
-        it('returns no validation errors', async () => {
-          expect(await validator.validateInput({ resource: httpOperations[0], input: GOOD_INPUT })).toEqual([]);
+        it('returns no validation errors', () => {
+          expect(validateInput({ resource: httpOperations[0], input: GOOD_INPUT })).toEqual([]);
         });
       });
     });
@@ -37,7 +34,7 @@ describe('HttpValidator', () => {
     describe('headers validation', () => {
       it('is case insensitive', async () => {
         expect(
-          await validator.validateInput({
+          await validateInput({
             resource: {
               method: 'GET',
               path: '/hey',
@@ -77,7 +74,7 @@ describe('HttpValidator', () => {
     describe('query validation', () => {
       it('returns only query validation errors', () => {
         expect(
-          validator.validateInput({
+          validateInput({
             resource: httpOperations[2],
             input: BAD_INPUT,
           }),
@@ -94,7 +91,7 @@ describe('HttpValidator', () => {
   describe('validateOutput()', () => {
     describe('all validations are turned on', () => {
       it('returns validation errors for whole request structure', async () => {
-        expect(await validator.validateOutput({ resource: httpOperations[1], output: BAD_OUTPUT })).toMatchSnapshot();
+        expect(await validateOutput({ resource: httpOperations[1], output: BAD_OUTPUT })).toMatchSnapshot();
       });
     });
   });
