+++ packages/http/src/validator/__tests__/index.spec.ts
@@ -1,10 +1,8 @@
 import { IPrismDiagnostic } from '@stoplight/prism-core';
-import { DiagnosticSeverity, IHttpContent, IHttpHeaderParam, IHttpOperation, IHttpQueryParam } from '@stoplight/types';
-import { IHttpNameValue, IHttpNameValues } from '../../types';
+import { DiagnosticSeverity, IHttpOperation } from '@stoplight/types';
 import { IHttpRequest } from '../../types';
-import { HttpValidator } from '../index';
+import { bodyValidator, headersValidator, queryValidator, validateInput, validateOutput } from '../index';
 import * as findResponseSpecModule from '../utils/spec';
-import { IHttpValidator } from '../validators/types';
 
 const mockError: IPrismDiagnostic = {
   message: 'mocked C is required',
@@ -14,24 +12,19 @@ const mockError: IPrismDiagnostic = {
 };
 
 describe('HttpValidator', () => {
-  const httpBodyValidator = { validate: () => [mockError] } as IHttpValidator<any, IHttpContent>;
-  const httpHeadersValidator = { validate: () => [mockError] } as IHttpValidator<IHttpNameValue, IHttpHeaderParam>;
-  const httpQueryValidator = { validate: () => [mockError] } as IHttpValidator<IHttpNameValues, IHttpQueryParam>;
-  const httpValidator = new HttpValidator(httpBodyValidator, httpHeadersValidator, httpQueryValidator);
-
   beforeEach(() => {
     jest.clearAllMocks();
     jest.spyOn(findResponseSpecModule, 'findOperationResponse').mockReturnValue(undefined);
-    jest.spyOn(httpBodyValidator, 'validate');
-    jest.spyOn(httpHeadersValidator, 'validate');
-    jest.spyOn(httpQueryValidator, 'validate');
+    jest.spyOn(bodyValidator, 'validate').mockReturnValue([mockError]);
+    jest.spyOn(headersValidator, 'validate').mockReturnValue([mockError]);
+    jest.spyOn(queryValidator, 'validate').mockReturnValue([mockError]);
   });
 
   describe('validateInput()', () => {
     describe('body validation in enabled', () => {
       const validate = (resourceExtension: Partial<IHttpOperation> | undefined, errorsNumber: number) => () => {
         expect(
-          httpValidator.validateInput({
+          validateInput({
             resource: Object.assign(
               {
                 method: 'get',
@@ -82,7 +75,7 @@ describe('HttpValidator', () => {
   describe('headers validation in enabled', () => {
     const validate = (resourceExtension?: Partial<IHttpOperation>, length: number = 1) => () => {
       expect(
-        httpValidator.validateInput({
+        validateInput({
           resource: Object.assign(
             {
               method: 'get',
@@ -110,7 +103,7 @@ describe('HttpValidator', () => {
       length: number = 2,
     ) => () => {
       expect(
-        httpValidator.validateInput({
+        validateInput({
           resource: Object.assign(
             {
               method: 'get',
@@ -125,9 +118,9 @@ describe('HttpValidator', () => {
         }),
       ).toHaveLength(length);
 
-      expect(httpBodyValidator.validate).not.toHaveBeenCalled();
-      expect(httpHeadersValidator.validate).toHaveBeenCalled();
-      expect(httpQueryValidator.validate).toHaveBeenCalledWith({}, [], undefined);
+      expect(bodyValidator.validate).not.toHaveBeenCalled();
+      expect(headersValidator.validate).toHaveBeenCalled();
+      expect(queryValidator.validate).toHaveBeenCalledWith({}, []);
     };
 
     describe('request is not set', () => {
@@ -153,7 +146,7 @@ describe('HttpValidator', () => {
     describe('output is set', () => {
       it('validates the body and headers', () => {
         expect(
-          httpValidator.validateOutput({
+          validateOutput({
             resource: {
               method: 'get',
               path: '/',
@@ -166,8 +159,8 @@ describe('HttpValidator', () => {
         ).toHaveLength(2);
 
         expect(findResponseSpecModule.findOperationResponse).toHaveBeenCalled();
-        expect(httpBodyValidator.validate).toHaveBeenCalledWith(undefined, [], undefined);
-        expect(httpHeadersValidator.validate).toHaveBeenCalled();
+        expect(bodyValidator.validate).toHaveBeenCalledWith(undefined, [], undefined);
+        expect(headersValidator.validate).toHaveBeenCalled();
       });
     });
   });
